backend/pyproject.toml
[project]
name = "soldieriq-backend"
version = "0.1.0"
description = "SoldierIQ - Tactical Intelligence Knowledge Management System Backend"
readme = "README.md"
requires-python = ">=3.11,<3.13"
authors = [
    { name = "SoldierIQ Team" }
]
dependencies = [
    # FastAPI and Server
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",
    # Pydantic
    "pydantic>=2.10.0",
    "pydantic-settings>=2.10.0",
    # Environment
    "python-dotenv>=1.1.0",
    # Database
    "pymongo>=4.9.0",
    "motor>=3.7.0",
    "redis>=6.2.0",
    # AI & LLM
    "openai>=1.99.0",
    "anthropic>=0.60.0",
    # Vector Database
    "pinecone>=5.0.0",
    "langchain>=0.2.7",
    "langchain-community>=0.2.6",
    "langchain-openai>=0.1.7",
    "langchain-pinecone>=0.1.3",
    # Document Processing
    "pypdfium2>=4.30.0",
    "pypdf>=5.9.0",
    "aiofiles>=24.1.0",
    "markitdown>=0.0.1a2",
    "unstructured-client>=0.29.0",
    "tiktoken>=0.8.0",
    # Web Scraping & YouTube
    "beautifulsoup4>=4.12.3",
    "html2text>=2025.4.0",
    "youtube-transcript-api>=1.2.1",
    # Cloud Storage
    "boto3>=1.40.0",
    "aioboto3>=13.0.0",
    # Utilities
    "pytz>=2025.2",
    "chonkie>=1.5.4",
]

